Refactor: MedigerList, Medicine 코드 수정 및 medigerList 디렉토리 추가
